.connect logo mobile responsive
